requirements.txt
-f /usr/share/pip-wheels
cachelib==0.4.1
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
cs50==8.0.1
Flask==2.0.2
Flask-Session==0.4.0
google-api-core==2.3.2
google-auth==2.3.3
google-cloud==0.34.0
google-cloud-vision==2.6.3
googleapis-common-protos==1.54.0
greenlet==1.1.2
grpcio==1.43.0
grpcio-status==1.43.0
idna==3.3
imutils==0.5.4
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
numpy==1.21.5
opencv-python==4.5.5.62
pandas==1.3.5
pdf2image==1.16.0
Pillow==8.4.0
proto-plus==1.19.8
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.26.0
rsa==4.8
six==1.16.0
SQLAlchemy==1.4.29
sqlparse==0.4.2
termcolor==1.1.0
urllib3==1.26.7
Werkzeug==2.0.2
